### 2023-02-06 60f0d01
build(pre-commit): update config
