+++ .pre-commit-config.yaml
@@ -3,23 +3,23 @@ repos:
   # Check commit messages on commit
   - repo: https://github.com/commitizen-tools/commitizen
     # Specify tag
-    rev: v2.37.1
+    rev: v2.40.0
     hooks:
       - id: commitizen
         stages: [commit-msg]
   - repo: https://github.com/python/black
-    rev: 22.10.0
+    rev: 23.1.0
     hooks:
       - id: black
         language_version: python3
       - id: black-jupyter
         language_version: python3
   - repo: https://github.com/asottile/blacken-docs
-    rev: v1.12.1
+    rev: 1.13.0
     hooks:
       - id: blacken-docs
   - repo: https://github.com/pycqa/isort
-    rev: 5.10.1
+    rev: 5.12.0
     hooks:
       - id: isort
         args: ["--profile", "black"]
@@ -31,7 +31,7 @@ repos:
           [flake8-docstrings, flake8-bugbear, flake8-bandit]
         files: ^fractopo/
   - repo: https://github.com/kynan/nbstripout
-    rev: 0.5.0
+    rev: 0.6.1
     hooks:
       - id: nbstripout
   - repo: https://github.com/pre-commit/pre-commit-hooks
