+++ tests/analysis/test_parameters.py
@@ -19,6 +19,7 @@ from fractopo.analysis import parameters
 #     reason="""
 # Odd bug with Windows and TKinter for only this single test.
 
+
 # See:
 #     """,
 # )
