### 2023-03-06 #5096 @elbeejay
Hi @nialov thank you for your submission and we appreciate your patience.

Do you have any suggestions for reviewers of your work? JOSS maintains a [list of potential reviewers](https://docs.google.com/spreadsheets/d/1PAPRJ63yq9aPC1COLjaQp8mHmEq3rZUzwUYxTulyu78/edit#gid=856801822), but we are also very receptive to any suggestions you may be have who have expertise in your field.

My only request at this time in terms of the paper/code, is that you reduce the entries in the `paper.bib` file to only include those references that are cited in the paper. I am not sure why our bot parses them all, but as you can see from the reference check comment above, there are a number of references with un-findable DOIs. At a glance it doesn't look like any of those are actually in the paper, so it'd be preferable if they were simply removed from the `.bib` file.

Thanks,
Jay
