### 2023-03-07 #5096 @nialov
Hey @elbeejay and thanks for picking up my submission!

Filtering the `.bib` file is no problem. I believe I already had tested a lua filter that filters based on what is actually used.
I will also post some suggestions for reviewers here. Will get these done in a day or two!
