### 2023-03-07 8918e25
docs(paper): use filtered paper.bib
Removed all unused referenced from paper.bib. Kept full
library of references in allrefs.bib for possible future use.
