+++ paper/refs.bib
@@ -1,385 +1,0 @@
-@article{nyberg_networkgt_2018,
-  author = {Nyberg, Björn and Nixon, Casey W. and Sanderson, David J.},
-  title = {{{NetworkGT}}: {{A}} {{GIS}} Tool for Geometric and
-    Topological Analysis of Two-Dimensional Fracture Networks},
-  journal = {Geosphere},
-  volume = {14},
-  number = {4},
-  pages = {1618-1634},
-  date = {2018-08},
-  urldate = {2021-11-24},
-  url = {https://pubs.geoscienceworld.org/gsa/geosphere/article/14/4/1618/531129/NetworkGT-A-GIS-tool-for-geometric-and-topological},
-  doi = {10.1130/GES01595.1},
-  issn = {1553-040X},
-  abstract = {Fractures rarely occur individually but more usually as
-    networks of numerous fractures whose arrangement, abundance, and
-    interaction control the mechanical and transport properties of rock
-    masses. Of particular importance are the distributions and spatial
-    variations of different geometric (locations, orientation, length,
-    etc.) and topological (intersections, connectivity, etc.)
-    ­attributes of fractures in a network. Geographical Information
-    Systems (GIS) provide a means to map and digitize two-dimensional
-    fracture networks from a variety of field and remote sensing data
-    and to display the results in the form of quality maps. We introduce
-    NetworkGT, an open-source toolbox for ArcGIS capable of efficient
-    sampling, analysis, and spatial mapping of geometric and topological
-    attributes of two-dimensional fracture networks. The toolbox helps
-    to extract and plot geometric and topological information from a
-    given two-dimensional fracture network including: rose diagrams,
-    plots of frequency distribution and topology, and maps of
-    topological parameters. Using a fracture network example from
-    offshore NW Devon, United Kingdom, we illustrate the practicality
-    and effectiveness of the toolbox. This includes computing a contour
-    grid with 1326 subsampled regions within the fracture network, which
-    is used to demonstrate the quantitative capabilities of the toolbox
-    and the ability to spatially map important network properties. The
-    toolbox will help to facilitate the increasing application of
-    geometry and topology in the analysis and comparison of fracture
-    networks at a range of scales. Furthermore, the integration of the
-    NetworkGT toolbox into ArcGIS allows two-dimensional fracture
-    networks to be interpreted, mapped, and fully analyzed within the
-    same software package.}
-}
-@article{skytta_fault-induced_2021,
-  author = {Skyttä, Pietari and Ovaskainen, Nikolas and Nordbäck,
-    Nicklas and Engström, Jon and Mattila, Jussi},
-  title = {Fault-Induced Mechanical Anisotropy and Its Effects on
-    Fracture Patterns in Crystalline Rocks},
-  journal = {Journal of Structural Geology},
-  volume = {146},
-  pages = {104304},
-  date = {2021-05},
-  urldate = {2021-11-24},
-  url = {https://linkinghub.elsevier.com/retrieve/pii/S0191814121000286},
-  doi = {10.1016/j.jsg.2021.104304},
-  issn = {01918141},
-  abstract = {We used drone-acquired orthophotographs to map the
-    2D-networks of fractures transecting the Mesoproterozoic Rapakivi
-    granites in southern Finland. The work aims at understanding i) how
-    discrete faults and the kine­ matically linked syn-fault extension
-    fractures originate in crystalline rocks, and ii) how the resulting
-    structural anisotropy controls the patterns of later-formed regional
-    fractures. We recognized incipient faults with associated distinct
-    damage zones (DZ; wall-, tip-, bend- and linking damage) indicating
-    that faulting occurred within mesoscopically isotropic material
[…]
