### 2023-03-07 #5096 @nialov
# Suggested Reviewers

Outside the spreadsheet:

-  Björn Nyberg (GitHub account: BjornNyberg)

   -  Author of <https://github.com/BjornNyberg/NetworkGT>, the "inspiration" for `fractopo`

-  Dennis Gläser (GitHub account: dglaeser)

   -  Author of `frackit` (<https://joss.theoj.org/papers/10.21105/joss.02291>) which
      resides in the same domain/topic as `fractopo`.

-  Casey Nixon (GitHub account: cwnixon)

   -  Structural geologist on whose work much of the functionality
      of `fractopo` is partly based on. Co-author of `NetworkGT`.

From the spreadsheet:

-  rowanc1

   -  Only one with "structural geology modelling" as a domain

Beyond these suggestions, I could not point out anyone specifically from
the spreadsheet. Anyone with a combination of GIS/spatial
analysis/geospatial/geoscience and (structural) geology would be
appropriate as far as I can tell.

`geopandas` and `shapely` are heavily used in my library so the people
familiar with them could probably by constructive (From spreadsheet:
martinfleis; outside spreadsheet: brendan-ward, jorisvandenbossche).

Hopefully helpful! Let me know if more suggestions are needed. I tried to filter to people at least somewhat familiar
with GitHub.
