### 2023-03-07 #5096 @elbeejay
@BjornNyberg and @dglaeser, would either of you be interested in reviewing this submission to JOSS titled: _"fractopo: A Python package for fracture network analysis"_? You have both developed similar tools, `NetworkGT` (@BjornNyberg) and `frackit` (@dglaeser) making you ideal reviewers for this submission.

At JOSS we do open checklist-driven reviews; peer-review criteria can be viewed [here](https://joss.readthedocs.io/en/latest/review_criteria.html). This issue is a "pre-review issue" which we use to find peer-reviewers. Once 2-3 reviewers are found, we will officially start the review in a dedicated GitHub issue. At present we are asking reviewers to complete reviews in 6 weeks, although this can be extended if needed. If you are not able to review but can recommend someone else, please mention them here (in this case please mention their GitHub handle without the "@" symbol).

If you are interested, please take a look at the journal's [conflict of interest policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html?#joss-conflict-of-interest-policy) to ensure you do not have a conflict before agreeing to review this submission.

Thanks again for taking a moment to consider this, feel free to contact me if you have any questions about the JOSS review process. Please do not feel any pressure to accept this review request if you do not have the time or do not feel comfortable reviewing this software package, we appreciate and respect our peer-reviewers' time. If you cannot serve as a reviewer at this time but have a peer, collaborator, student, or colleague who might be available and would be a good fit for this submission, please let me know! If interested we can figure out how to set up a "co-review" for a colleague of yours that is a more junior or inexperienced member of the community.

Thanks,
Jay
