### 2023-03-24 #5096 @ICWallis
Good luck with your PhD Fernanda. Last push!

Sent from a phone (typos and autocorrect included for your amusement)

> On 24/03/2023, at 4:38 PM, Fernanda Alvarado-Neves ***@***.***> wrote:
>
> Fernanda
