### 2023-03-24 #5300 @editorialbot
[REVIEW]: fractopo: A Python package for fracture network analysis
**Submitting author:** @nialov (<a href="http://orcid.org/0000-0003-1562-0280">Nikolas Ovaskainen</a>)
**Repository:** https://github.com/nialov/fractopo
**Branch with paper.md** (empty if default branch):
**Version:** v0.5.3
