### 2023-03-24 #5300 @elbeejay
Thanks again @dglaeser and @lachlangrose for agreeing to review `fractopo`. Instructions about how to generate your reviewer checklist as well as how to conduct a JOSS review are provided in the above comments, but please feel free to reach out to me if you have any questions. I'll be asking the bot to send out reminders about the review in 3 weeks time; note that we are currently asking our reviewers to complete their reviews within 6 weeks.
