### 2023-04-12 9d99987
fix: handle z-coordinates in geometries
Should solve basic problems associated with z-coordinates
in shapely geometries such as the one posted in #21.
