### 2023-04-12 2f61e3b
test: add tests with z-coordinate input data
