+++ tests/__init__.py
@@ -337,106 +337,11 @@ geta_lidar_inf_valid_area = read_geofile(
 )
 
 
-# def random_data_column(iterable):
-#     """
-#     Make random data column contents.
-#     """
-#     return ["aaa" for _ in iterable]
-
-
-# geoms are all LineStrings and no errors
-
-
-# def valid_gdf_get():
-#     """
-#     Get valid gdf.
-#     """
-#     return gpd.GeoDataFrame(
-#         {
-#             GEOMETRY_COLUMN: valid_traces,
-#             ERROR_COLUMN: valid_error_srs,
-#             "random_col": random_data_column(valid_traces),
-#             "random_col2": random_data_column(valid_traces),
-#             "random_col3": random_data_column(valid_traces),
-#             "random_col4": random_data_column(valid_traces),
-#         }
-#     )
-
-
-# def invalid_gdf_get():
-#     """
-#     Get invalid gdf.
-#     """
-#     return gpd.GeoDataFrame(
-#         {
-#             GEOMETRY_COLUMN: invalid_traces,
-#             ERROR_COLUMN: invalid_error_srs,
-#             "random_col": random_data_column(invalid_traces),
-#         }
-#     )
-
-
-# def invalid_gdf_null_get():
-#     """
-#     Get gdf with None and empty geometries.
-#     """
-#     return gpd.GeoDataFrame(
-#         {
-#             GEOMETRY_COLUMN: [None, LineString()],
-#             ERROR_COLUMN: [[], []],
-#             "random_col": random_data_column(range(2)),
-#         }
-#     )
-
-
-# def valid_area_gdf_get():
-#     """
-#     Get a valid area gdf.
-#     """
-#     return gpd.GeoDataFrame({GEOMETRY_COLUMN: valid_areas_geoseries})
-
-
-# def invalid_area_gdf_get():
-#     """
-#     Get an invalid area gdf.
-#     """
-#     return gpd.GeoDataFrame({GEOMETRY_COLUMN: invalid_areas_geoseries})
-
-
 faulty_error_srs = pd.Series([[] for _ in valid_traces.geometry.values])
 faulty_error_srs[0] = np.nan
 faulty_error_srs[1] = "this cannot be transformed to list?"
 faulty_error_srs[2] = (1, 2, 3, "hello?")
 faulty_error_srs[5] = 5.12315235
-
-
-# def valid_gdf_with_faulty_error_col_get():
-#     """
-#     Get valid gdf with faulty error column.
-#     """
-#     return gpd.GeoDataFrame(
-#         {
-#             GEOMETRY_COLUMN: valid_traces,
-#             ERROR_COLUMN: faulty_error_srs,
-#             "random_col": random_data_column(valid_traces),
-#         }
-#     )
-
-
-# def iterate_validators():
-#     """
-#     Iterate over validators.
-#     """
-#     for validator in (
-#         GeomNullValidator,
-#         GeomTypeValidator,
-#         MultiJunctionValidator,
-#         VNodeValidator,
-#         MultipleCrosscutValidator,
-#         UnderlappingSnapValidator,
-#         TargetAreaSnapValidator,
-#     ):
-#         yield validator
 
 
 nice_integer_coordinates = integers(-10, 10)
@@ -495,20 +400,6 @@ def get_nice_traces():
     Get nice traces GeoSeries.
     """
     return nice_traces.copy()
-
-
-# def get_traces_geosrs():
-#     """
-#     Get traces GeoSeries.
-#     """
-#     return traces_geosrs.copy()
-
-
-# def get_areas_geosrs():
-#     """
-#     Get areas GeoSeries.
-#     """
-#     return areas_geosrs.copy()
 
 
[…]
