+++ tests/analysis/test_network/test_network_traces2_area2_KB11_0_100_True_True_0_001_False_True_True_.yml
@@ -1,0 +1,94 @@
+Area: 1237.9
+Areal Frequency B20: 0.22
+Areal Frequency P20: 0.07
+Branch Boundary 0 Intersect Count: 282
+Branch Boundary 1 Intersect Count: 0
+Branch Boundary 2 Intersect Count: 0
+Branch Max Length: 15.69
+Branch Mean Length: 1.07
+Branch Min Length: 0.01
+C - C: 190
+C - E: 28
+C - I: 64
+Censoring: .nan
+Connection Frequency: 0.12
+Connections per Branch: 1.76
+Connections per Trace: 3.37
+Dimensionless Intensity B22: 0.25
+Dimensionless Intensity P22: 0.66
+E: 28
+E - E: 0
+Fracture Density (Mauldon): .nan
+Fracture Intensity (Mauldon): .nan
+Fracture Intensity B21: 0.23
+Fracture Intensity P21: 0.23
+I: 64
+I - E: 0
+I - I: 0
+Number of Branches: 268.0
+Number of Branches (Real): 282
+Number of Traces: 86.0
+Number of Traces (Real): 100
+Radius: .nan
+Relative Censoring: .nan
+Trace Boundary 0 Intersect Count: 100
+Trace Boundary 1 Intersect Count: 0
+Trace Boundary 2 Intersect Count: 0
+Trace Max Length: 24.57
+Trace Mean Length: 2.87
+Trace Mean Length (Mauldon): .nan
+Trace Min Length: 0.1
+X: 37
+Y: 108
+branch exponential Kolmogorov-Smirnov distance D: 0.21
+branch exponential lambda: 0.51
+branch exponential loglikelihood: -78.44
+branch lengths cut off proportion: 0.83
+branch lognormal Kolmogorov-Smirnov distance D: .nan
+branch lognormal loglikelihood: -69.23
+branch lognormal mu: -874.02
+branch lognormal sigma: 22.77
+branch lognormal vs. exponential R: 2.64
+branch lognormal vs. exponential p: 0.01
+branch power_law Kolmogorov-Smirnov distance D: 0.06
+branch power_law alpha: 2.69
+branch power_law cut-off: 1.5
+branch power_law exponent: -1.69
+branch power_law sigma: 0.25
+branch power_law vs. exponential R: 2.63
+branch power_law vs. exponential p: 0.01
+branch power_law vs. lognormal R: -0.26
+branch power_law vs. lognormal p: 0.79
+branch power_law vs. truncated_power_law R: -0.3
+branch power_law vs. truncated_power_law p: 0.79
+branch truncated_power_law Kolmogorov-Smirnov distance D: 0.06
+branch truncated_power_law alpha: 2.61
+branch truncated_power_law exponent: -1.61
+branch truncated_power_law lambda: 0.01
+branch truncated_power_law loglikelihood: -69.24
+trace exponential Kolmogorov-Smirnov distance D: 0.17
+trace exponential lambda: 0.27
+trace exponential loglikelihood: -76.39
+trace lengths cut off proportion: 0.67
+trace lognormal Kolmogorov-Smirnov distance D: 0.06
+trace lognormal loglikelihood: -72.77
+trace lognormal mu: -4.82
+trace lognormal sigma: 2.09
+trace lognormal vs. exponential R: 1.56
+trace lognormal vs. exponential p: 0.12
+trace power_law Kolmogorov-Smirnov distance D: 0.08
+trace power_law alpha: 2.61
+trace power_law cut-off: 2.88
+trace power_law exponent: -1.61
+trace power_law sigma: 0.28
+trace power_law vs. exponential R: 1.36
+trace power_law vs. exponential p: 0.17
+trace power_law vs. lognormal R: -0.3
[…]

[…]
