### 2023-04-12 6991ad6
docs: add information on Z-coordinates
This, along with previous commits should close #21.
