### 2023-04-12 08a0ada
docs(misc): add info on snap_threshold and crs
Rephrased the info that was posted in #22.
