+++ noxfile.py
@@ -256,7 +256,6 @@ def _docs(session, auto_build: bool):
     auto_build_args = (
         [
             f"--ignore=**/{DOCS_AUTO_EXAMPLES_PATH.name}/**",
-            "--ignore=**/.fractopo_cache/**",
             "--watch=README.rst",
             f"--watch={PACKAGE_NAME}/",
             "--watch=examples/",
