### 2023-04-15 nialov/fractopo#25 @lachlangrose
Error when plotting trace lengths
Hi,

I'm just starting my review of the code for JOSS.

I'm using a google colab environment and have installed fractopo using `pip install git+https://github.com/nialov/fractopo` - so this should be the most recent version of the code.

I have tried to run your example https://fractopo.readthedocs.io/en/latest/notebooks/fractopo_network_1.html but I am getting a matplotlib error.

```
[<ipython-input-12-a892d99a2aba>](https://localhost:8080/#) in <cell line: 1>()
----> 1 fit, fig, ax = kb11_network.plot_trace_lengths()

5 frames
[/usr/local/lib/python3.9/dist-packages/matplotlib/axis.py](https://localhost:8080/#) in set_ticks(self, ticks, labels, minor, **kwargs)
   2070         """
   2071         if labels is None and kwargs:
-> 2072             raise ValueError('labels argument cannot be None when '
   2073                              'kwargs are passed')
   2074         result = self._set_tick_locations(ticks, minor=minor)

ValueError: labels argument cannot be None when kwargs are passed
```

I'm using matplotlib 3.7.1. and have a gist here if you want to try running it https://gist.github.com/lachlangrose/506df31fdd81019df4a7e1d5126cac40
