### 2023-04-15 nialov/fractopo#26 @lachlangrose
Review question: Ignoring warnings
Hi,

In the documentation the recommendation is to ignore warnings. I think this is potentially a bad idea as the warnings are usually there for a reason and it could have unintended consequences for other libraries that may be used in the same workflow. Would it be possible to filter the specific warnings which you deem at unncessary?
