### 2023-04-15 nialov/fractopo#26 @lachlangrose
Following on from this point, I have seen there are a lot of warnings from parts of this code which would not be difficult to remove.

For example to add another dataframe to an existing dataframe:

`df.append(addition, ignore_index=True)` can be replaced with `df = df.concat([df,addition],ignore_index=True)`

to do the same thing multiple times for example in ` determine_intersects` you can build up a list of intersecting dataframes and then using a python list and then concat all of these lists together once the loop has been run.

This should speed up the code and will also make it more future proof. Another warning `ShapelyDeprecationWarning: The array interface is deprecated and will no longer work in Shapely 2.0. Convert the '.coords' to a numpy array instead.
  arr = construct_1d_object_array_from_listlike(values)` would be worth fixing as well.
