### 2023-04-15 nialov/fractopo#25 @nialov
Thank you for the report! Based on short testing it it related to the newer `matplotlob` version. 3.6.3 is the `matplotlib` version used in development that is locked in `poetry.lock`. I will issue a fix of either constraining the `matplotlib` version in `pyproject.toml` that `pip` uses for installs or fixing the actual code to support the newer version.

Until I have issued a fix you can specify the version like so in the notebook:

~~~
!pip install git+https://github.com/nialov/fractopo matplotlib==3.6.3
~~~
