### 2023-04-15 3af4493
fix(analysis): fix length plot ticks
Refactored the tick setting code to explicitly pass the labels as well
as the tick locations. This fixes #25.
