### 2023-04-15 ced54eb
build(poetry): update matplotlib to newer version
Also constrained the minimum and maximum versions in pyproject.toml.
