+++ pyproject.toml
@@ -39,7 +39,7 @@ python = ">=3.8,<3.12"
 click = "*"
 geopandas = ">=0.11.1,<1.0.0"
 joblib = ">=1.2.0,<2.0.0"
-matplotlib = "*"
+matplotlib = ">=3.6.3,<4.0.0"
 numpy = "*"
 pandas = ">=1.3,<2.0.0"
 powerlaw = "*"
