+++ environment.yml
@@ -8,7 +8,7 @@ dependencies:
   - python >=3.8,<3.12
   - click
   - geopandas >=0.11.1,<1.0.0
-  - matplotlib-base
+  - matplotlib-base >=3.63,<4.0.0
   - numpy
   - pandas >=1.3,<2.0.0
   - powerlaw
