### 2023-04-15 d466e54
build(environment): fix matplotlib-base spec
