### 2023-04-15 47159ae
build(pyproject): allow matplotlib 3.5.3
