### 2023-04-15 9dc6444
docs(notebooks): specify ignore warnings
Rather than recommending to ignore all warnings, the warnings that have
been found to be mostly harmless are now specified (#26).
