### 2023-04-15 a288a97
test: use strict checking for warnings in pytest
Now all expected warnings, which are mostly related to underlying
dependency code, are ignore explicitly while all new warnings will cause
an error to be raised during pytest execution.
