+++ pyproject.toml
@@ -126,15 +126,13 @@ doctest_optionflags =[
     ]
 
 filterwarnings =[
-    "ignore:invalid value encountered in true_divide:RuntimeWarning",
-    "ignore:The Shapely GEOS version :UserWarning",
-    "ignore:invalid value encountered in double_scalars:RuntimeWarning",
-    "ignore:Mean of empty slice:RuntimeWarning",
-    "ignore:Degrees of freedom :RuntimeWarning",
-    "ignore:divide by zero encountered in double_scalars:RuntimeWarning",
-    "ignore:This method will be removed in future versions:DeprecationWarning",
-    "ignore:The default dtype for empty Series will be:DeprecationWarning",
-    "ignore:All-NaN slice encountered:RuntimeWarning",
+    "error",
+    "ignore:The Shapely:UserWarning:geopandas",
+    "ignore:No data for colormapping provided:UserWarning",
+    "ignore:invalid value encountered in divide:RuntimeWarning:powerlaw",
+    "ignore:The behavior of :FutureWarning",
+    "ignore:In a future version:FutureWarning",
+    "ignore:invalid value encountered in intersection:RuntimeWarning:pygeos",
     ]
 
 [tool.isort]
