### 2023-04-15 88b40ee
fix(length_distribution): handle empty array
Previously empty arrays could be passed to powerlaw.Fit
but now they are handled before downstream processing. The
output from determining a fit for empty data is explicitly now None.
