### 2023-04-15 ed13a2b
refactor(parameters): handle invalid data cases
Empty plots are now returned early and logarithmic
axes are not used on invalid data.
