### 2023-04-15 cb07f73
test: fix minor issues causing warnings
