### 2023-04-15 c7799c4
test(cli): remove doctest
The functions that output to stdout do not work that well in doctests.
