+++ fractopo/cli.py
@@ -74,14 +74,6 @@ def describe_results(
 ):
     """
     Describe validation results to stdout.
-
-    E.g.
-
-    >>> gdf = gpd.GeoDataFrame({"VALIDATION_ERRORS": [("V NODE",)]})
-    >>> describe_results(gdf, "VALIDATION_ERRORS")
-    Out of 1 traces, 1 were invalid.
-    There were 1 error types. These were:
-    V NODE
     """
     error_count = sum(len(val) != 0 for val in validated[error_column].values)
     error_types = {
