### 2023-04-15 2e1fa73
fix(branches_and_nodes): fix small bugs
These were causing warnings. Using the DataFrame was anyway overkill for
the task in additional_snapping_func.
