### 2023-04-15 d2c9e36
refactor(relationships): use pandas concat
Replaced append with concat and speeded up code.
