### 2023-04-15 24aa2de
refactor(relationships): use prepared.prep
STRTree causes warnings so switched to another implementation of fast
spatial queries.
