### 2023-04-15 4532ea9
refactor: warnings should be passed to user
Warnings from underlying dependencies should not be ignored without a
good reason when fractopo is used as a library.
