+++ docs_src/requirements.txt
@@ -50,7 +50,7 @@ identify==2.5.13 ; python_version >= "3.8" and python_version < "3.12"
 idna==3.4 ; python_version >= "3.8" and python_version < "3.12"
 imagesize==1.4.1 ; python_version >= "3.8" and python_version < "3.12"
 importlib-metadata==6.0.0 ; python_version >= "3.8" and python_version < "3.12"
-importlib-resources==5.10.2 ; python_version >= "3.8" and python_version < "3.9"
+importlib-resources==5.10.2 ; python_version >= "3.8" and python_version < "3.10"
 iniconfig==2.0.0 ; python_version >= "3.8" and python_version < "3.12"
 ipykernel==6.20.2 ; python_version >= "3.8" and python_version < "3.12"
 ipython-genutils==0.2.0 ; python_version >= "3.8" and python_version < "3.12"
@@ -77,7 +77,7 @@ kiwisolver==1.4.4 ; python_version >= "3.8" and python_version < "3.12"
 livereload==2.6.3 ; python_version >= "3.8" and python_version < "3.12"
 markupsafe==2.1.2 ; python_version >= "3.8" and python_version < "3.12"
 matplotlib-inline==0.1.6 ; python_version >= "3.8" and python_version < "3.12"
-matplotlib==3.6.3 ; python_version >= "3.8" and python_version < "3.12"
+matplotlib==3.7.1 ; python_version >= "3.8" and python_version < "3.12"
 mistune==2.0.4 ; python_version >= "3.8" and python_version < "3.12"
 mpmath==1.2.1 ; python_version >= "3.8" and python_version < "3.12"
 munch==2.5.0 ; python_version >= "3.8" and python_version < "3.12"
