### 2023-04-16 70183e7
ci(conda): use libmamba solver
This speeds up the workflow.
