+++ .github/workflows/conda.yml
@@ -51,6 +51,8 @@ jobs:
           environment-file: environment.yml
           python-version: ${{ matrix.python-version }}
           auto-activate-base: false
+          # Use libmamba solver for faster CI
+          mamba-version: "*"
           # IMPORTANT: This needs to be set for caching to work properly!
           use-only-tar-bz2: true
       - name: Print conda environment
