### 2023-04-16 882937d
build(pyproject): improve warning specification
