+++ pyproject.toml
@@ -128,10 +128,11 @@ doctest_optionflags =[
 filterwarnings =[
     "error",
     "ignore:The Shapely:UserWarning:geopandas",
+    "ignore:The shapely GEOS version:UserWarning:pygeos",
     "ignore:No data for colormapping provided:UserWarning",
     "ignore:invalid value encountered in divide:RuntimeWarning:powerlaw",
     "ignore:The behavior of :FutureWarning",
-    "ignore:In a future version:FutureWarning",
+    "ignore:In a future version:FutureWarning:geopandas",
     "ignore:invalid value encountered in intersection:RuntimeWarning:pygeos",
     ]
 
