### 2023-04-16 9b32f48
fix(branches_and_nodes): fix angle_to_point
Fixed the distance and similarity checks between points to more
reasonable accuracy.
