### 2023-04-16 f3c6573
test: disable joblib cache in pytest execution
This makes tests less flaky although specific flakyness has not been
detected so far. But now disk caching is not tested during the test runs
by default.
