### 2023-04-16 d0a4faf
build(pyproject): improve warning list
