+++ pyproject.toml
@@ -132,8 +132,9 @@ filterwarnings =[
     "ignore:No data for colormapping provided:UserWarning",
     "ignore:invalid value encountered in divide:RuntimeWarning:powerlaw",
     "ignore:The behavior of :FutureWarning",
-    "ignore:In a future version:FutureWarning:geopandas",
+    "ignore:In a future version, :FutureWarning",
     "ignore:invalid value encountered in intersection:RuntimeWarning:pygeos",
+    "ignore:More than 20 figures have been opened:RuntimeWarning",
     ]
 
 [tool.isort]
