### 2023-04-16 7c9cbde
test: ignore matplotlib warning upstream
