### 2023-04-16 nialov/fractopo#25 @nialov
Pull request #27 was merged which should fix this specific issue. The notebook you included in your gist now works. If the same issue persists on your end you can report it here! I will close this after a week or two otherwise.
