### 2023-04-16 bb8e0d6
build(pyproject): add powerlaw warning
