### 2023-04-16 a8caa81
test: handle TclError raised in Windows ci
