### 2023-04-16 4292c17
refactor(analysis): specify optional Fit
