### 2023-04-17 nialov/fractopo#26 @nialov
Thank you for your feedback. The warning ignoring has been made more explicit
in our documentation, i.e. the notebooks, to avoid any unintended consequences.
Additionally, we have implemented numerous code fixes, including all the ones
you mentioned in your post, in pull request #28.

Furthermore, we have implemented a more strict approach regarding warnings. The
expected errors are now better explicitly listed in `pyproject.toml`. This was
also the case previously but your issue report prompted a general debugging and
cleaning session where I hunted where warnings came from to assess if they were
mostly harmless or related to underlying dependencies in which they were added
to `pyproject.toml`.

The debugging turned into a general bug hunting session for me so do not be
worried about the amount of commits that prompted from this issue report :D. It
is good to check through these more minor warnings every now and then to keep
the code base "more future proof" as you also pointed out.
