### 2023-04-17 721ea1b
test(analysis): label more tests with decorator
