### 2023-04-17 5af3daf
test(analysis): fix decorator logic
