### 2023-04-25 nialov/fractopo#29 @dglaeser
Minor issues in the readme
I was reading through the readme file and stumbled across a few (minor) things. I'll post them here in case you want to consider them:

-  The readme says "As it is a Python library, the use of fractopo requires prior (Python) programming knowledge.". I have the feeling that the CLI may already be useful for people, so maybe I would rephrase or remove this sentence not to scare off some people.
- (this is actually not so minor) When I tried the instructions using a local installation with pip from the sources with `pip install --editable .[dev]`, I got a warning that the extra `dev` is not defined. This actually causes a bunch of follow-up errors. Because if you then try to run `poetry run pytest`, one gets a long list of errors from missing modules... I think the pip install should either be fixed, or, the instructions removed, requiring users to rely on `poetry`
- In the analysis example (https://github.com/nialov/fractopo#geometric-and-topological-trace-network-analysis), it would be good if the comment above the `snap_threshold` stated the unit in which this is given?
- The `development` section (https://github.com/nialov/fractopo#geometric-and-topological-trace-network-analysis) starts with "Breaking changes are possible and expected." This doesn't sound super inviting for people that want to use the package :). Since you have a changelog, maybe one can add that any breaking changes will be properly documented therein?
