### 2023-04-25 969a0f9
docs(readme): use interface instead of script
