### 2023-04-25 bbe56f4
docs(readme): reword use recommendations
As pointed out in #29, the command-line use of fractopo
should be possible without Python knowledge.
