### 2023-04-25 7e5d6e6
docs(readme): specify development install method
Specified that only the poetry installation is supported. Supporting pip
can cause issues during installation (e.g. #29) and due to the
dependencies not being the same as with poetry installation.
