### 2023-04-25 aac781b
docs(readme): indicate snap threshold unit
