### 2023-04-25 92beee4
docs(readme): reword breaking changes info
