### 2023-04-25 #5300 @elbeejay
Hi @nialov, @dglaeser, and @lachlangrose

I wanted to check-in here and ask how the review process is going. Both @dglaeser and @lachlangrose have opened up issues in the repository - @nialov how is the resolution to those items going?

@dglaeser and @lachlangrose have you finished your initial reviews of the software?

Thanks all,
Jay
