### 2023-04-25 nialov/fractopo#31 @dglaeser
Link validation errors in README
I think it could be nice to add a link to the docs about validation errors (https://fractopo.readthedocs.io/en/latest/validation/errors.html) directly in the repo readme in the section `trace validation`, such that the reader can directly get an impression what `fractopo` checks there.
