### 2023-04-25 nialov/fractopo#32 @dglaeser
Add context to validation error docs
Regarding the validation error docs: are these errors physically motivated (i.e. to detect non-physical topologies) or are they there for some other technical reason? I guess some are technical, some are physically-motivated. Do you think it makes sense to say a word or two on that in the respective error sections? Some actually contain very useful background information, for instance, the `MultiLineString` error.

For the sharp-corner-validator, the docs say

"... is completely open to interpretation and therefore the resulting errors are subjective."

Wouldn't a similar argument also hold for other errors? For instance, under- or overlapping snap? For instance, the underlapping-snap-docs say

"Underlapping error can occur when a trace ends very close to another trace but not near enough. The abutting might not be registered as a Y-node."

In this context, "not near enough" means it's further away than the given `snap_threshold`? If this is correct, then I guess the threshold would be the subjectivity? But on the other hand, I then wonder why it should be an error? I am probably missing something...

_Originally posted by @dglaeser in https://github.com/nialov/fractopo/issues/31#issuecomment-1521957012_
