### 2023-04-25 nialov/fractopo#32 @dglaeser
A follow-up question on that: `VNodeValidator` seems to me like a special case of `MultiJunctionValidator` (the docs mention that "V NODE errors might also be detected as MULTI JUNCTION errors"). Is the former a common case and therefore is worth keeping as a special validation/error?

As a side note: there seems to be an issue with backticks and a missing whitespace in the sentence "V NODE errors might also be detected asMULTI JUNCTION` errors"
