### 2023-04-25 nialov/fractopo#34 @dglaeser
[CLI][tracevalidate] save validation results into a sub-dir of `pwd`?
As far as I see it, when invoking `fractopo tracevalidate PATH_TO_TRACES PATH_TO_AREA`, the default is that the validated data is saved in a sub-directory of the input data (e.g. PATH_TO_TRACES)?

That may be intentional, but I was wondering if it is more intuitive to put that data into sub-directories of the current working dir per default?
