### 2023-04-25 nialov/fractopo#35 @dglaeser
[CLI][tracevalidate] default value for `allow-fix`
`fractopo tracevalidate --help` shows for the `allow-fix` option

```bash
--allow-fix,--fix                                             Allow the direct modification of trace file to fix errors. [default: True]
```

Does this really modify the file (i.e. resaving it to disk)? If so, I am unsure if `True` should be the default. Modification of the input file should be explicitly requested I would say. Or does this rather mean that the errors are fixed and then the fixed version of the traces are saved into `output`?  if the latter is the case, maybe the docstring for this CLI parameter could be adjusted a bit?
