### 2023-04-25 nialov/fractopo#36 @dglaeser
[CLI][tracevalidate] make area a named argument?
Currently, the CLI invocation is `fractopo tracevalidate TRACES AREA` as far as I see it. While this is quite intuitive, in the very beginning I had to double check which one comes first. I also then tried swapping the two, and I seem to recall that this didn't lead to an error message that makes it directly obvious what went wrong. Therefore, I was wondering if `tracevalidate` should receive just the traces as a positional argument, and the area as one that is passed with a name, e.g. `fractopo tracevalidate traces.geojson --area my_area.geojson`.
