### 2023-04-25 nialov/fractopo#29 @dglaeser
I also noticed that the link to the `CONTRIBUTING.rst` file in the `development` section of the readme seems broken?
