### 2023-04-25 #5300 @dglaeser
@elbeejay, I opened a few (rather minor) issues:

- [ ] https://github.com/nialov/fractopo/issues/29
- [ ] https://github.com/nialov/fractopo/issues/31
- [ ] https://github.com/nialov/fractopo/issues/32
- [ ] https://github.com/nialov/fractopo/issues/33
- [ ] https://github.com/nialov/fractopo/issues/34
- [ ] https://github.com/nialov/fractopo/issues/35
- [ ] https://github.com/nialov/fractopo/issues/36

I will have one more pass over the paper tomorrow and tick off the final marks and/or post remarks about the paper.
