### 2023-04-26 #5300 @nialov
@elbeejay All issue reports so far have been very helpful in making `fractopo` a better package both in terms of code and documentation! I have already fixed issues by both dglaeser and lachlangrose in pull requests and will continue to do so in the next few weeks. Thank you for the reviews so far!
