### 2023-04-26 #5300 @dglaeser
@elbeejay, I added some more points to my previous comment with the related issues/prs. I just left one tickbox open in my review, for which an open issue exists. My points raised are actually rather minor and also partly subjective. Thanks @nialov for this very nice package.
