### 2023-04-26 88da872
docs(readme): remove fix issues from #29
Removed inconsistent pip section and fixed link to
CONTRIBUTING.rst
