### 2023-04-26 c28e8e5
docs(readme): add link to validation errors
Part of suggestions in #32.
