### 2023-04-26 8e467b9
docs(docs_src): fix links in usage table
Fixes #33.
