### 2023-04-26 396ade4
docs(errors): add context to errors
As suggested in #32, context has been added on most error types.
This includes the motivation, either technical or physical.
