### 2023-04-26 b993e42
refactor(cli): make error type print conditional
