### 2023-04-26 62b0eb6
refactor(cli): save validated to local directory
As suggested in #34, the validated trace data is now saved to a
directory relative to the current directory if the output is not
explicitly set.
