### 2023-04-26 ee6fb40
refactor(cli): adjust cli help for allow_fix
It is now mentioned that the input file is not modified unless
explicitly set as the output even with the --allow-fix flag
as suggested in #35.
