### 2023-04-26 337736a
feat(cli): check input trace and area geometries
This should help in cases where the user accidentally passes the traces
and area in the wrong order. The check will cause the invocation to fail
fast if the geometries are not what is expected. This should help with
issue #36.
