### 2023-04-27 nialov/fractopo#32 @dglaeser
I think this should be fixed in #40? Will close this now, thanks a lot!
