### 2023-04-27 nialov/fractopo#36 @dglaeser
With #41, one now gets notified if the order is wrong, so I'll close this, thanks!
