### 2023-04-27 7ad5411
docs(paper): lay out summary better
Attempt to fix problems reported in #38 related to
the end of the summary.
