### 2023-04-27 #5300 @nialov
Hey, @elbeejay! All open issues by @dglaeser are handled. There are currently two open issues by @lachlangrose:

-  https://github.com/nialov/fractopo/issues/25
-  https://github.com/nialov/fractopo/issues/26

Both for which pull request have been made and merged. I am now waiting on responses to the fixes and for any more issues to fix!
