### 2023-05-03 #5300 @elbeejay
@nialov I had a few small suggestions for the text:

Line 10-13: I suggest revising this sentence to something like "Digitizing these fracture traces from observations (e.g., drone imaged outcrops) results in georeferenced two-dimensional fracture trace vector datasets (i.e., fracture networks), which offer..." to help make it clear what text the "e.g.," and "i.e.," apply to.

Line 35: Consider a small revision to reword "boundary to fractopo based on which a number of analysis results..." to something like "boundary to fractopo, from which a number of analysis results..."

Line 38: Is the "e.g." needed here? If these are all examples I suggest simply saying something like "Example results (Figure 2) include rose plots ... "

Line 47-48: A minor text suggestion "package managers as well as containing features absent..." could be revised to something like "package managers and contains features absent..."

Once you have completed any revisions and finalized the manuscript text, please do the following:

- [ ] Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) (in this case please update the archived version)
- [ ] (please consider) Re-releasing the package on PyPi and conda so that there are versions in the package managers that correspond to what is going to be the JOSS-publication version
