### 2023-05-09 nialov/fractopo#46 @nialov
Editor suggestions for the manuscript
From: https://github.com/openjournals/joss-reviews/issues/5300

Line 10-13: I suggest revising this sentence to something like "Digitizing these fracture traces from observations (e.g., drone imaged outcrops) results in georeferenced two-dimensional fracture trace vector datasets (i.e., fracture networks), which offer..." to help make it clear what text the "e.g.," and "i.e.," apply to.

Line 35: Consider a small revision to reword "boundary to fractopo based on which a number of analysis results..." to something like "boundary to fractopo, from which a number of analysis results..."

Line 38: Is the "e.g." needed here? If these are all examples I suggest simply saying something like "Example results (Figure 2) include rose plots ... "

Line 47-48: A minor text suggestion "package managers as well as containing features absent..." could be revised to something like "package managers and contains features absent..."

_Originally posted by @elbeejay in https://github.com/openjournals/joss-reviews/issues/5300#issuecomment-1533507023_
