### 2023-05-09 21dce9e
docs(paper): implement suggested changes
Implements all changes suggested in #46. Furthermore, contains small
fixes and rephrasing.
