### 2023-05-09 5e93dad
docs(paper): remove 'in'
