+++ CITATION.cff
@@ -6,8 +6,8 @@ authors:
   given-names: Nikolas
   orcid: "https://orcid.org/0000-0003-1562-0280"
 title: "fractopo"
-version: "0.5.2"
-date-released: "2023-01-16"
+version: "0.5.3"
+date-released: "2023-05-09"
 keywords:
   - python
   - data
