### 2023-05-10 #5300 @nialov
All done @elbeejay

> Once you have completed any revisions and finalized the manuscript text, please do the following:
>
>     * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.

Version tag is `v0.5.3`.

>
>     * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository) (in this case please update the archived version)
>
>     * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
>     * [ ]  Please list the DOI of the archived version here.

DOI is https://doi.org/10.5281/zenodo.7915808

>
>     * [ ]  (please consider) Re-releasing the package on PyPi and conda so that there are versions in the package managers that correspond to what is going to be the JOSS-publication version

PyPi package has been released. Conda will follow shortly as the feedstock workflow finishes: <https://github.com/conda-forge/fractopo-feedstock>
